+++ src/Resources/contao/languages/en/default.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the ContaoMmenuBundle.
+ *
+ * (c) Dirk Klemmt
+ * (c) inspiredminds
+ *
+ * @license MIT
+ */
+
+$GLOBALS['TL_LANG']['DK_MMENU']['placeholder'] = 'Search';
+$GLOBALS['TL_LANG']['DK_MMENU']['noResult'] = 'No results found.';
+$GLOBALS['TL_LANG']['DK_MMENU']['title'] = 'Menu';