+++ src/Resources/public/mmenu/addons/counters/mmenu.counters.js
@@ -1,0 +1,66 @@
+import Mmenu from '../../core/oncanvas/mmenu.oncanvas';
+import options from './_options';
+import { extendShorthandOptions } from './_options';
+import * as DOM from '../../_modules/dom';
+import { extend } from '../../_modules/helpers';
+//	Add the options.
+Mmenu.options.counters = options;
+//	Add the classnames.
+Mmenu.configs.classNames.counters = {
+    counter: 'Counter'
+};
+export default function () {
+    var _this = this;
+    var options = extendShorthandOptions(this.opts.counters);
+    this.opts.counters = extend(options, Mmenu.options.counters);
+    //	Refactor counter class
+    this.bind('initListview:after', function (panel) {
+        var cntrclss = _this.conf.classNames.counters.counter, counters = panel.querySelectorAll('.' + cntrclss);
+        counters.forEach(function (counter) {
+            DOM.reClass(counter, cntrclss, 'mm-counter');
+        });
+    });
+    //	Add the counters after a listview is initiated.
+    if (options.add) {
+        this.bind('initListview:after', function (panel) {
+            if (!panel.matches(options.addTo)) {
+                return;
+            }
+            var parent = panel['mmParent'];
+            if (parent) {
+                //	Check if no counter already excists.
+                if (!parent.querySelector('.mm-counter')) {
+                    var counter = DOM.create('span.mm-counter');
+                    var btn = DOM.children(parent, '.mm-btn')[0];
+                    if (btn) {
+                        btn.prepend(counter);
+                    }
+                }
+            }
+        });
+    }
+    if (options.count) {
+        var count = function (panel) {
+            var panels = panel
+                ? [panel]
+                : DOM.children(_this.node.pnls, '.mm-panel');
+            panels.forEach(function (panel) {
+                var parent = panel['mmParent'];
+                if (!parent) {
+                    return;
+                }
+                var counter = parent.querySelector('.mm-counter');
+                if (!counter) {
+                    return;
+                }
+                var listitems = [];
+                DOM.children(panel, '.mm-listview').forEach(function (listview) {
+                    listitems.push.apply(listitems, DOM.children(listview));
+                });
+                counter.innerHTML = DOM.filterLI(listitems).length.toString();
+            });
+        };
+        this.bind('initListview:after', count);
+        this.bind('updateListview', count);
+    }
+}