+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+### 3.1.0
+
+* Updated mmenu to 8.3.0
+* Added columns as option
+* Added mmenu IE10 & IE11 Polyfill as option 
+
+### 3.0.0
+
+* Updated mmenu to 8.1.1
+
 ### 1.5.1, 2014-12-23
 * Updated mmenu to version 4.7.5
 * Added russian translation (Thanks Dmitri Kuznetsov! :smile:)